+++ CHANGELOG.md
@@ -74,6 +74,7 @@
 * New cop `SelfAssignment` checks for places where the self-assignment shorthand should have been used. ([@bbatsov][])
 * New cop `DoubleNegation` checks for uses of `!!`. ([@bbatsov][])
 * New cop `PercentLiteralDelimiters` enforces consistent usage of `%`-literal delimiters. ([@hannestyden][])
+* New cop `MultilineAssignment` enforces multi-line values to new line and indentation and indentation. ([@hannestyden][])
 * New Rails cop `ActionFilter` enforces the use of `_filter` or `_action` action filter methods. ([@bbatsov][])
 * New Rails cop `ScopeArgs` makes sure you invoke the `scope` method properly. ([@bbatsov][])
 * Add `with_fixed_indentation` style to `AlignParameters` cop. ([@hannestyden][])